+++ ingeniamotion/communication.py
@@ -34,6 +34,8 @@ if TYPE_CHECKING:
 from ingeniamotion.metaclass import DEFAULT_AXIS, DEFAULT_SERVO, MCMetaClass
 
 RUNNING_ON_WINDOWS = platform.system() == "Windows"
+FILE_EXT_SFU = ".sfu"
+FILE_EXT_LFU = ".lfu"
 
 
 class Communication(metaclass=MCMetaClass):
@@ -999,7 +1001,28 @@ class Communication(metaclass=MCMetaClass):
                 error_enabled_callback,
             )
 
-    def load_firmware_ecat(self, ifname: str, fw_file: str, slave: int = 1) -> None:
+    @staticmethod
+    def __get_boot_in_app(fw_file: str) -> bool:
+        """Return true if the bootloader is included in the application (file extension is .sfu).
+
+        Args:
+            fw_file: Path to the FW file.
+
+        Returns:
+            True if the bootloader is included in the application.
+
+        Raises:
+            ValueError: If the firmware file has the wrong extension.
+        """
+        if not fw_file.endswith((FILE_EXT_SFU, FILE_EXT_LFU)):
+            raise ValueError(
+                f"Firmware file should have extension {FILE_EXT_SFU} or {FILE_EXT_LFU}."
+            )
+        return fw_file.endswith(FILE_EXT_SFU)
+
+    def load_firmware_ecat(
+        self, ifname: str, fw_file: str, slave: int = 1, boot_in_app: Optional[bool] = None
+    ) -> None:
         """Load firmware via ECAT.
 
         Args:
@@ -1007,9 +1030,12 @@ class Communication(metaclass=MCMetaClass):
                 ``\\Device\\NPF_[...]``.
             fw_file : Firmware file path.
             slave : slave index. ``1`` by default.
+            boot_in_app: true if the bootloader is included in the application, false otherwise.
+                If None, the file extension is used to define it.
 
         Raises:
             FileNotFoundError: If the firmware file cannot be found.
+            ValueError: If the firmware file has the wrong extension.
             ingenialink.exceptions.ILFirmwareLoadError: If no slave is detected.
             ingenialink.exceptions.ILFirmwareLoadError: If the FoE write operation is not successful
             NotImplementedError: If FoE is not implemented for the current OS and architecture
@@ -1018,9 +1044,10 @@ class Communication(metaclass=MCMetaClass):
 
         net = EthercatNetwork(ifname)
         if fw_file.endswith(self.ENSEMBLE_FIRMWARE_EXTENSION):
-            self.__load_ensemble_fw_ecat(net, fw_file, slave)
+            self.__load_ensemble_fw_ecat(net, fw_file, slave, boot_in_app)
         else:
-            net.load_firmware(fw_file, slave)
+            boot_in_app = self.__get_boot_in_app(fw_file) if boot_in_app is None else boot_in_app
+            net.load_firmware(fw_file, boot_in_app, slave)
 
     def load_firmware_ecat_interface_index(
         self, if_index: int, fw_file: str, slave: int = 1
@@ -1243,7 +1270,9 @@ class Communication(metaclass=MCMetaClass):
                 f"Load of FW in slave {slave_id} of ensemble failed. Exception: {exception}"
             )
 
-    def __load_ensemble_fw_ecat(self, net: EthercatNetwork, fw_file: str, slave: int) -> None:
+    def __load_ensemble_fw_ecat(
+        self, net: EthercatNetwork, fw_file: str, slave: int, boot_in_app: Optional[bool]
+    ) -> None:
         """Load FW to an ensemble of servos through Ethercat.
 
         The FW files that are contained in the fw_file will be loaded by selecting any slave which
@@ -1253,13 +1282,24 @@ class Communication(metaclass=MCMetaClass):
             net: Ethercat network.
             fw_file: Path to the ensemble FW file.
             slave: Slave ID (any slave in the ensemble)
+            boot_in_app: true if the bootloader is included in the application, false otherwise.
+                If None, the file extension is used to define it.
         """
         mapping = self.__unzip_ensemble_fw_file(fw_file)
         scanned_slaves = net.scan_slaves_info()
         first_slave_in_ensemble = self.__check_ensemble(scanned_slaves, slave, mapping)
         try:
             for slave_id_offset, fw_file_prod_code in mapping.items():
-                net.load_firmware(fw_file_prod_code[0], first_slave_in_ensemble + slave_id_offset)
+                boot_in_app_drive = (
+                    self.__get_boot_in_app(fw_file_prod_code[0])
+                    if boot_in_app is None
+                    else boot_in_app
+                )
+                net.load_firmware(
+                    fw_file_prod_code[0],
+                    boot_in_app_drive,
+                    first_slave_in_ensemble + slave_id_offset,
+                )
         except ILError as e:
             raise e
         finally: