+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- Set the boot_in_app flag according to the file extension.
 - FSoE module.
 
 ### Fix